[google-cloud-jupyter-config] Fix async test coverage and bugs this exposed.
The previous version of the async test case did not actually run to the end
because we never awaited the result. That meant that we weren't actually
testing the async behavior the way we thought we were and there were hidden
bugs in it that were not caught by our tests.

This change fixes the test so that async test cases actually run end to end,
and also fixes the preexisting bugs that were caught by the fixed tests.